moves asset importers & tasks into an "assets" namespace. Adds rake import:assets:all
