+++ lib/importers/assets/other_deposits_importer.rb
@@ -1,0 +1,25 @@
+require 'importers/base_importer'
+
+module Importers
+  module Assets
+    class OtherDepositsImporter < BaseImporter
+      def import!
+        each_row do |row|
+          person = Person.find_by!(internal_code: row[:codigopersona])
+          declaration = person.assets_declarations.first!
+
+          kind           = row[:clase]
+          description    = row[:descripcion]
+          amount         = parse_amount(row[:numero_cuantia_o_valor_en_euros])
+          purchase_date  = parse_spanish_date(row[:fecha_de_adquisicion])
+
+          unless declaration.has_other_deposit?(kind, description, amount, purchase_date)
+            puts "Importing other deposit for #{person.name} (#{kind}, #{description}, #{amount})"
+            declaration.add_other_deposit(kind, description, amount, purchase_date)
+            declaration.save!
+          end
+        end
+      end
+    end
+  end
+end
